Split up test_transport
Remove further unnecessary arguments
